+++ src/components/apps/infra-viewer/InfraViewer.vue
@@ -9,7 +9,6 @@
 import Vue from 'vue'
 import SchemaPicker from './SchemaPicker.vue';
 import {GraphSeriesOption} from 'echarts'
-import {getLinks, getNodes } from './crs-to-series';
 
 export default Vue.extend({
   name: 'InfraViewer',
